# [processor/lsminterval]Introduce metric overflow limit