--- a/processor/lsmintervalprocessor/processor.go
+++ b/processor/lsmintervalprocessor/processor.go
@@ -102,7 +102,8 @@ func newProcessor(cfg *config.Config, ivlDefs []intervalDef, log *zap.Logger, ne
 				v := merger.NewValue(
 					cfg.ResourceLimit,
 					cfg.ScopeLimit,
-					cfg.ScopeDatapointLimit,
+					cfg.MetricLimit,
+					cfg.DatapointLimit,
 				)
 				if err := v.Unmarshal(value); err != nil {
 					return nil, fmt.Errorf("failed to unmarshal value from db: %w", err)
@@ -111,7 +112,8 @@ func newProcessor(cfg *config.Config, ivlDefs []intervalDef, log *zap.Logger, ne
 					v,
 					cfg.ResourceLimit,
 					cfg.ScopeLimit,
-					cfg.ScopeDatapointLimit,
+					cfg.MetricLimit,
+					cfg.DatapointLimit,
 				), nil
 			},
 		},
@@ -254,7 +256,8 @@ func (p *Processor) ConsumeMetrics(ctx context.Context, md pmetric.Metrics) erro
 	v := merger.NewValue(
 		p.cfg.ResourceLimit,
 		p.cfg.ScopeLimit,
-		p.cfg.ScopeDatapointLimit,
+		p.cfg.MetricLimit,
+		p.cfg.DatapointLimit,
 	)
 	md.ResourceMetrics().RemoveIf(func(rm pmetric.ResourceMetrics) bool {
 		rm.ScopeMetrics().RemoveIf(func(sm pmetric.ScopeMetrics) bool {
@@ -418,7 +421,8 @@ func (p *Processor) exportForInterval(
 		v := merger.NewValue(
 			p.cfg.ResourceLimit,
 			p.cfg.ScopeLimit,
-			p.cfg.ScopeDatapointLimit,
+			p.cfg.MetricLimit,
+			p.cfg.DatapointLimit,
 		)
 		if err := v.Unmarshal(iter.Value()); err != nil {
 			errs = append(errs, fmt.Errorf("failed to decode binary from database: %w", err))