--- a/processor/lsmintervalprocessor/processor_test.go
+++ b/processor/lsmintervalprocessor/processor_test.go
@@ -113,9 +113,10 @@ func TestAggregationOverflow(t *testing.T) {
 					},
 				},
 			},
-			ResourceLimit:       oneCardinalityLimitConfig,
-			ScopeLimit:          oneCardinalityLimitConfig,
-			ScopeDatapointLimit: oneCardinalityLimitConfig,
+			ResourceLimit:  oneCardinalityLimitConfig,
+			ScopeLimit:     oneCardinalityLimitConfig,
+			MetricLimit:    oneCardinalityLimitConfig,
+			DatapointLimit: oneCardinalityLimitConfig,
 		}
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()