Introduces a metric overflow limit and removes the previous scope datapoint limit. Now, the limits follow the same hierarchy as the pmetric datastructure:

1. Resource limit
2. Scope limit
3. Metric limit
4. Datapoint limit

It is also possible to only set some of the limits, for example: if we only set datapoint limit then there will only be limit on the datapoint per metric and rest all will be without limit.

Closes: https://github.com/elastic/opentelemetry-collector-components/issues/141