@lahsivjar 2025-01-14
Benchmark with main results in ~5% increase in allocation (probably because of extra allocations required for trackers for metric limits):

```
name                                            old time/op    new time/op    delta
Aggregation/sum_cumulative-10                     12.3µs ± 5%    12.2µs ± 5%    ~     (p=0.902 n=7+7)
Aggregation/sum_delta-10                          12.2µs ± 1%    12.3µs ± 2%    ~     (p=0.209 n=7+7)
Aggregation/histogram_cumulative-10               12.9µs ± 5%    12.7µs ± 2%    ~     (p=0.534 n=7+6)
Aggregation/histogram_delta-10                    12.8µs ± 4%    13.0µs ± 2%    ~     (p=0.318 n=7+7)
Aggregation/exphistogram_cumulative-10            12.1µs ± 2%    12.7µs ± 3%  +4.39%  (p=0.002 n=7+7)
Aggregation/exphistogram_delta-10                 13.1µs ± 6%    12.9µs ± 3%    ~     (p=0.945 n=7+6)
Aggregation/summary_enabled-10                    13.1µs ±15%    12.3µs ± 2%  -6.65%  (p=0.026 n=7+7)
Aggregation/summary_passthrough-10                1.34µs ± 7%    1.30µs ± 1%    ~     (p=0.093 n=7+5)
AggregationWithOTTL/sum_cumulative-10             13.7µs ±10%    13.2µs ± 3%    ~     (p=0.535 n=7+7)
AggregationWithOTTL/sum_delta-10                  13.2µs ± 2%    13.1µs ± 4%    ~     (p=0.620 n=7+7)
AggregationWithOTTL/histogram_cumulative-10       14.1µs ± 3%    13.9µs ± 5%    ~     (p=0.181 n=7+6)
AggregationWithOTTL/histogram_delta-10            14.1µs ± 5%    13.9µs ± 3%    ~     (p=0.805 n=7+7)
AggregationWithOTTL/exphistogram_cumulative-10    14.6µs ± 3%    14.2µs ± 2%    ~     (p=0.097 n=7+7)
AggregationWithOTTL/exphistogram_delta-10         14.5µs ± 4%    14.3µs ± 2%    ~     (p=0.383 n=7+7)
AggregationWithOTTL/summary_enabled-10            13.8µs ± 3%    13.9µs ± 2%    ~     (p=0.534 n=7+6)
AggregationWithOTTL/summary_passthrough-10        1.34µs ± 5%    1.29µs ± 2%  -3.60%  (p=0.035 n=7+6)

name                                            old alloc/op   new alloc/op   delta
Aggregation/sum_cumulative-10                     15.0kB ± 1%    15.6kB ± 1%  +4.09%  (p=0.001 n=7+6)
Aggregation/sum_delta-10                          14.9kB ± 0%    15.6kB ± 1%  +4.84%  (p=0.003 n=5+7)
Aggregation/histogram_cumulative-10               20.6kB ± 6%    20.3kB ± 1%    ~     (p=0.710 n=7+7)
Aggregation/histogram_delta-10                    20.1kB ± 6%    20.2kB ± 1%    ~     (p=0.209 n=7+7)
Aggregation/exphistogram_cumulative-10            15.4kB ± 0%    16.1kB ± 0%  +4.32%  (p=0.001 n=7+7)
Aggregation/exphistogram_delta-10                 15.4kB ± 0%    16.0kB ± 0%  +4.08%  (p=0.001 n=7+6)
Aggregation/summary_enabled-10                    13.7kB ± 2%    14.0kB ± 0%  +2.61%  (p=0.006 n=6+7)
Aggregation/summary_passthrough-10                1.37kB ± 1%    1.37kB ± 1%    ~     (p=0.877 n=7+7)
AggregationWithOTTL/sum_cumulative-10             15.4kB ± 1%    15.9kB ± 1%  +3.26%  (p=0.001 n=7+7)
AggregationWithOTTL/sum_delta-10                  15.3kB ± 0%    15.9kB ± 0%  +3.33%  (p=0.001 n=6+7)
AggregationWithOTTL/histogram_cumulative-10       17.6kB ± 0%    18.1kB ± 0%  +2.79%  (p=0.001 n=7+6)
AggregationWithOTTL/histogram_delta-10            17.6kB ± 1%    18.1kB ± 0%  +3.22%  (p=0.001 n=7+7)
AggregationWithOTTL/exphistogram_cumulative-10    17.7kB ± 0%    18.2kB ± 1%  +3.25%  (p=0.001 n=7+7)
AggregationWithOTTL/exphistogram_delta-10         17.7kB ± 1%    18.3kB ± 1%  +3.14%  (p=0.001 n=7+7)
AggregationWithOTTL/summary_enabled-10            15.7kB ± 0%    16.3kB ± 1%  +3.69%  (p=0.001 n=7+7)
AggregationWithOTTL/summary_passthrough-10        1.37kB ± 1%    1.37kB ± 2%    ~     (p=0.245 n=7+7)

name                                            old allocs/op  new allocs/op  delta
Aggregation/sum_cumulative-10                        147 ± 0%       155 ± 0%  +5.44%  (p=0.001 n=7+7)
Aggregation/sum_delta-10                             148 ± 0%       156 ± 0%  +5.41%  (p=0.001 n=7+7)
Aggregation/histogram_cumulative-10                  209 ± 1%       220 ± 1%  +5.26%  (p=0.001 n=7+7)
Aggregation/histogram_delta-10                       211 ± 1%       221 ± 1%  +4.67%  (p=0.001 n=7+7)
Aggregation/exphistogram_cumulative-10               165 ± 0%       173 ± 0%  +4.85%  (p=0.001 n=7+7)
Aggregation/exphistogram_delta-10                    166 ± 0%       174 ± 0%  +4.82%  (p=0.001 n=7+7)
Aggregation/summary_enabled-10                       156 ± 0%       164 ± 0%  +5.13%  (p=0.001 n=7+7)
Aggregation/summary_passthrough-10                  35.4 ± 2%      35.6 ± 2%    ~     (p=1.000 n=7+7)
AggregationWithOTTL/sum_cumulative-10                174 ± 0%       182 ± 0%  +4.60%  (p=0.001 n=7+7)
AggregationWithOTTL/sum_delta-10                     175 ± 0%       183 ± 0%  +4.57%  (p=0.001 n=7+7)
AggregationWithOTTL/histogram_cumulative-10          188 ± 0%       196 ± 0%  +4.26%  (p=0.001 n=7+6)
AggregationWithOTTL/histogram_delta-10               189 ± 0%       197 ± 0%  +4.23%  (p=0.001 n=7+7)
AggregationWithOTTL/exphistogram_cumulative-10       192 ± 0%       200 ± 0%  +4.17%  (p=0.001 n=7+7)
AggregationWithOTTL/exphistogram_delta-10            193 ± 0%       201 ± 0%  +4.15%  (p=0.001 n=7+7)
AggregationWithOTTL/summary_enabled-10               183 ± 0%       191 ± 0%  +4.37%  (p=0.001 n=7+7)
AggregationWithOTTL/summary_passthrough-10          35.6 ± 2%      35.3 ± 2%    ~     (p=0.592 n=7+7)
```